pyproject.toml
[project]
name = "wellinformed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "biopython>=1.86",
    "docling>=2.61.1",
    "fastmcp>=2.13.0.2",
    "google-adk>=1.18.0",
    "google-cloud-firestore>=2.21.0",
    "google-genai>=1.49.0",
    "google-search-results>=2.4.2",
    "habanero>=2.3.0",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "matplotlib>=3.10.7",
    "mem0ai>=1.0.0",
    "networkx>=3.5",
    "pandas>=2.3.3",
    "plotly>=6.4.0",
    "pydantic>=2.12.4",
    "pyeuropepmc>=1.8.1",
    "python-dotenv>=1.2.1",
    "pytrials>=1.0.0",
    "requests>=2.32.5",
    "semanticscholar>=0.11.0",
    "tavily-python>=0.7.12",
    "toolbox-core>=0.5.2",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
    "xmltodict>=1.0.2",
]

[tool.ruff]
line-length = 88
exclude = ["kaggle_google_course"]

[tool.ruff.lint]
ignore = [
    "E402",  # Module level import not at top of file - required for __future__ imports and sys.path setup
]